# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Zoophilist Pet Services API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: string;
  customerName: string;
  customerPhone: string;
  customerEmail?: string;
  city?: string;
  area?: string;
  address?: string;
  petName: string;
  petType: string;
  breed?: string;
  age?: string;
  aggressive?: boolean;
  serviceId?: string;
  serviceName: string;
  preferredDate?: string;
  preferredTime?: string;
  notes?: string;
  photoUrls?: string[];
  videoUrls?: string[];
  status: BookingStatus;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBookingInput {
  customerName: string;
  customerPhone: string;
  customerEmail?: string;
  city?: string;
  area?: string;
  address?: string;
  petName: string;
  petType: string;
  breed?: string;
  age?: string;
  aggressive?: boolean;
  serviceId: string;
  serviceName: string;
  preferredDate?: string;
  preferredTime?: string;
  notes?: string;
  photoUrls?: string[];
  videoUrls?: string[];
}

export type UpdateBookingStatusInputStatus = typeof UpdateBookingStatusInputStatus[keyof typeof UpdateBookingStatusInputStatus];


export const UpdateBookingStatusInputStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface UpdateBookingStatusInput {
  status: UpdateBookingStatusInputStatus;
}

export interface BookingListResponse {
  bookings: Booking[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface ServiceStat {
  serviceName: string;
  count: number;
  percentage: number;
}

export interface BookingStats {
  totalRequests: number;
  pending: number;
  confirmed: number;
  completed: number;
  cancelled: number;
  totalRevenue: number;
  weeklyRequests: number;
  popularServices: ServiceStat[];
  recentBookings: Booking[];
}

export interface Service {
  id: string;
  name: string;
  price: number;
  originalPrice?: number;
  description?: string;
  features: string[];
  isSubscription?: boolean;
  badge?: string;
  imageUrl?: string;
  createdAt?: string;
}

export interface CreateServiceInput {
  name: string;
  price: number;
  originalPrice?: number;
  description?: string;
  features: string[];
  isSubscription?: boolean;
  badge?: string;
  imageUrl?: string;
}

export interface UpdateServiceInput {
  name?: string;
  price?: number;
  originalPrice?: number;
  description?: string;
  features?: string[];
  isSubscription?: boolean;
  badge?: string;
  imageUrl?: string;
}

export type GalleryItemType = typeof GalleryItemType[keyof typeof GalleryItemType];


export const GalleryItemType = {
  image: 'image',
  video: 'video',
} as const;

export interface GalleryItem {
  id: string;
  url: string;
  type: GalleryItemType;
  category: string;
  caption?: string;
  createdAt: string;
}

export type CreateGalleryItemInputType = typeof CreateGalleryItemInputType[keyof typeof CreateGalleryItemInputType];


export const CreateGalleryItemInputType = {
  image: 'image',
  video: 'video',
} as const;

export interface CreateGalleryItemInput {
  url: string;
  type: CreateGalleryItemInputType;
  category: string;
  caption?: string;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminSession {
  id: string;
  username: string;
}

export interface AdminAuthResult {
  token: string;
  admin: AdminSession;
}

export type GetBookingsParams = {
status?: GetBookingsStatus;
page?: number;
limit?: number;
};

export type GetBookingsStatus = typeof GetBookingsStatus[keyof typeof GetBookingsStatus];


export const GetBookingsStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type GetGalleryParams = {
category?: GetGalleryCategory;
};

export type GetGalleryCategory = typeof GetGalleryCategory[keyof typeof GetGalleryCategory];


export const GetGalleryCategory = {
  all: 'all',
  grooming: 'grooming',
  bath: 'bath',
  haircut: 'haircut',
  happypets: 'happypets',
} as const;