# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Zoophilist Pet Services API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: string;
  customerName: string;
  customerPhone: string;
  customerEmail?: string;
  city?: string;
  area?: string;
  address?: string;
  petName: string;
  petType: string;
  breed?: string;
  age?: string;
  aggressive?: boolean;
  serviceId?: string;
  serviceName: string;
  preferredDate?: string;
  preferredTime?: string;
  notes?: string;
  photoUrls?: string[];
  videoUrls?: string[];
  status: BookingStatus;
  createdAt: string;
  updatedAt: string;
}