# lib/api-zod/src/generated/types/bookingStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Zoophilist Pet Services API
 * OpenAPI spec version: 0.1.0
 */
import type { Booking } from './booking';
import type { ServiceStat } from './serviceStat';

export interface BookingStats {
  totalRequests: number;
  pending: number;
  confirmed: number;
  completed: number;
  cancelled: number;
  totalRevenue: number;
  weeklyRequests: number;
  popularServices: ServiceStat[];
  recentBookings: Booking[];
}