# lib/api-zod/src/generated/types/createGalleryItemInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Zoophilist Pet Services API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateGalleryItemInputType } from './createGalleryItemInputType';

export interface CreateGalleryItemInput {
  url: string;
  type: CreateGalleryItemInputType;
  category: string;
  caption?: string;
}