# lib/api-zod/src/generated/types/serviceStat.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Zoophilist Pet Services API
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceStat {
  serviceName: string;
  count: number;
  percentage: number;
}