# lib/api-zod/src/generated/types/updateBookingStatusInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Zoophilist Pet Services API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBookingStatusInputStatus } from './updateBookingStatusInputStatus';

export interface UpdateBookingStatusInput {
  status: UpdateBookingStatusInputStatus;
}