# lib/api-zod/src/generated/types/updateServiceInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Zoophilist Pet Services API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateServiceInput {
  name?: string;
  price?: number;
  originalPrice?: number;
  description?: string;
  features?: string[];
  isSubscription?: boolean;
  badge?: string;
  imageUrl?: string;
}